+++ src/f200.h
@@ -58,8 +58,8 @@ namespace rsimpl
         void on_before_start(const std::vector<subdevice_mode_selection> & selected_modes) override;
         rs_stream select_key_stream(const std::vector<rsimpl::subdevice_mode_selection> & selected_modes) override;
         
-        void set_options(const rs_option options[], int count, const double values[]) override;
-        void get_options(const rs_option options[], int count, double values[]) override;
+        void set_options(const rs_option options[], size_t count, const double values[]) override;
+        void get_options(const rs_option options[], size_t count, double values[]) override;
 
         std::shared_ptr<frame_timestamp_reader> create_frame_timestamp_reader() const override;
     };
